GCI-1072: Outline new order confirmation class.
