+++ src/main/java/uk/gov/companieshouse/itemhandler/email/CertifiedDocument.java
@@ -1,40 +1,3 @@
 package uk.gov.companieshouse.itemhandler.email;
 
-public class CertifiedDocument {
-    private String dateFiled;
-    private String type;
-    private String description;
-    private String fee;
-
-    public String getDateFiled() {
-        return dateFiled;
-    }
-
-    public void setDateFiled(String dateFiled) {
-        this.dateFiled = dateFiled;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getFee() {
-        return fee;
-    }
-
-    public void setFee(String fee) {
-        this.fee = fee;
-    }
-}
+public class CertifiedDocument extends FilingHistoryDocument {}
