+++ src/main/java/uk/gov/companieshouse/itemhandler/email/FilingHistoryDocument.java
@@ -1,0 +1,40 @@
+package uk.gov.companieshouse.itemhandler.email;
+
+public class FilingHistoryDocument {
+    private String dateFiled;
+    private String type;
+    private String description;
+    private String fee;
+
+    public String getDateFiled() {
+        return dateFiled;
+    }
+
+    public void setDateFiled(String dateFiled) {
+        this.dateFiled = dateFiled;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getFee() {
+        return fee;
+    }
+
+    public void setFee(String fee) {
+        this.fee = fee;
+    }
+}
