+++ src/main/java/uk/gov/companieshouse/itemhandler/email/MissingImage.java
@@ -1,0 +1,3 @@
+package uk.gov.companieshouse.itemhandler.email;
+
+public class MissingImage extends FilingHistoryDocument {}
