+++ src/main/java/uk/gov/companieshouse/itemhandler/email/MissingImageDeliveryOrderConfirmation.java
@@ -1,0 +1,13 @@
+package uk.gov.companieshouse.itemhandler.email;
+
+public class MissingImageDeliveryOrderConfirmation extends OrderConfirmation {
+    private MissingImage missingImage;
+
+    public MissingImage getMissingImage() {
+        return missingImage;
+    }
+
+    public void setMissingImage(MissingImage missingImage) {
+        this.missingImage = missingImage;
+    }
+}
